requirements-mac.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apache-beam==2.61.0
apache-flink==2.0.0
apache-flink-libraries==2.0.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
avro==1.12.0
certifi==2025.4.26
charset-normalizer==3.4.2
cloudpickle==2.2.1
crcmod==1.7
dataclasses-json==0.6.7
dill==0.3.1.1
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docopt==0.6.2
dotenv==0.9.9
elastic-transport==8.17.1
elasticsearch==8.17.1
exceptiongroup==1.3.0
fastavro==1.11.1
fasteners==0.19
feedparser==6.0.11
filelock==3.18.0
find_libpython==0.4.0
frozenlist==1.6.0
fsspec==2025.5.0
greenlet==3.2.2
grpcio==1.65.5
h11==0.16.0
hdfs==2.7.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpickle==3.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.10
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.1.3
objsize==0.7.1
ollama==0.4.8
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pemja==0.4.1
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==1.4.2
pyflink==1.0
PyJWT==2.9.0
pymongo==4.13.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.25.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.32.0
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
